Employer jobs page displays correctly
