+++ controllers/index.js
@@ -1,11 +1,13 @@
-const router = require('express').Router();
+const router = require("express").Router();
 
 // const categoryRoutes = require('./categories');
-const homeRoutes = require('./homeRoutes');
-const apiRoutes = require('./api');
+const homeRoutes = require("./homeRoutes");
+const apiRoutes = require("./api");
+const profileRoutes = require("./profileRoutes");
 
-router.use('/', homeRoutes);
-router.use('/api', apiRoutes);
+router.use("/", homeRoutes);
+router.use("/api", apiRoutes);
+router.use("/profile", profileRoutes);
 // router.use('/category', categoryRoutes);
 
 module.exports = router;
