+++ views/employerProfile.handlebars
@@ -6,7 +6,7 @@
     <div class="row mt-5 pt-2">
       <div class="d-flex flex-row freelancerprofileheader pt-5">
         <h3 class="">Employer Profile Page</h3>
-        <a href="/employer/jobs" class="btn btn-primary btn-sm h-10 ms-auto mb-3">Create job listing</a>
+        <a href="/profile/employer/jobs" class="btn btn-primary btn-sm h-10 ms-auto mb-3">Create job listing</a>
       </div>
       <div class="col-lg-4 col-md-4 col-sm-12">
         <div>
